+++ weaver/wps_restapi/quotation/utils.py
@@ -1,13 +1,18 @@
 import logging
 from typing import TYPE_CHECKING
 
+from pyramid.httpexceptions import HTTPNotFound
 from pyramid.settings import asbool
 
+from weaver.database import get_db
 from weaver.config import WeaverFeature, get_weaver_configuration
+from weaver.exceptions import QuoteNotFound
+from weaver.store.base import StoreQuotes
 from weaver.utils import get_settings
 
 if TYPE_CHECKING:
-    from weaver.typedefs import AnySettingsContainer
+    from weaver.datatype import Quote
+    from weaver.typedefs import AnyRequestType, AnySettingsContainer
 
 
 LOGGER = logging.getLogger(__name__)
@@ -32,3 +37,21 @@ def check_quotation_supported(container):
         LOGGER.info(f"Unsupported quotation requests for configuration '%s'.", weaver_config)
         return False
     return True
+
+
+def get_quote(request):
+    # type: (AnyRequestType) -> Quote
+    """
+    Obtain the referenced :term:`Quote` by the request with validation.
+
+    :param request:
+    :return: Matched quote.
+    :raises HTTPNotFound: If the quote could not be found.
+    """
+    quote_id = request.matchdict.get("quote_id")
+    store = get_db(request).get_store(StoreQuotes)
+    try:
+        quote = store.fetch_by_id(quote_id)
+    except QuoteNotFound:
+        raise HTTPNotFound("Could not find quote with specified 'quote_id'.")
+    return quote